Fixed some bugs.

+++ controllers/cartController.js
@@ -7,7 +7,7 @@ const addCartItem = async (req,res)=>{
 
         const cartItem = new Cart({product, quantity, userId})
     
-        const cart = await cartItem.save()
+        await cartItem.save()
         res.status(201).send("Item added")
     } catch (error) {
         console.log(error)
@@ -20,8 +20,8 @@ const addCartItem = async (req,res)=>{
 const findByuserId = async (req,res) =>{
     try {
         const {id} = req.query
-        const cartItems = await Cart.find({userId:id}).exec()
-        if (!cartItems){
+        const cartItems = await Cart.find({userId:id}).populate("product") .populate("userId").exec()
+        if (!cartItems || cartItems.length == 0){
             return res.status(404).send("Could not fetch cart items")
         }
     
@@ -36,8 +36,8 @@ const findByuserId = async (req,res) =>{
 const deleteById = async (req,res)=>{
     try {
         const {id} = req.query
-        const deletedCount = await Cart.deleteOne({_id:id}).exec()
-        if(!deletedCount){
+        const result = await Cart.deleteOne({_id:id}).exec()
+        if(result.deletedCount === 0){
             return res.status(404).send("Could not be deleted")
         }
         res.status(200).send("Deleted successfully")
@@ -54,12 +54,12 @@ const updateById = async (req,res)=>{
 
         const data = req.body
     
-        const updated = await Cart.updateOne({_id:id},data)
-        if(!updated){
+        const result = await Cart.updateOne({_id:id},data)
+        if(result.matchedCount===0){
             return res.status(404).send("Could not be updated")
         }
     
-        res.staus(200).send("Updated successfully")
+        res.status(200).send("Updated successfully")
     } catch (error) {
         console.log(error)
         res.status(500).send("Server side error")
@@ -72,13 +72,17 @@ const increaseCartItemQuantity = async (req,res)=>{
         const {id} = req.query
 
         const cartItem = await Cart.findOne({_id:id})
-    
+        
+        if (!cartItem) {
+            return res.status(404).send("Cart item not found")
+        }
+
         const quantity = cartItem.quantity+1
     
         const data = {quantity:quantity}
     
-        const updated = await Cart.updateOne({_id:id},data)
-        if(!updated){
+        const result = await Cart.updateOne({_id:id},data)
+        if(result.matchedCount===0){
             return res.status(404).send("Couldnot be updated")
         }
         res.status(200).send("Updated sucessfully")
@@ -97,19 +101,22 @@ const decreaseCartItemQuantity = async(req,res)=>{
 
         const cartItem = await Cart.findOne({_id:id})
     
+        if (!cartItem) {
+            return res.status(404).send("Cart item not found")
+        }
     
         const quantity = cartItem.quantity-1
     
         if (quantity == 0){
-            const result = await Cart.deleteOne({_id:id})
+            await Cart.deleteOne({_id:id})
             console.log("Item deleted as quantity reached 0")
             return res.status(200).send("Item removed from cart")
         }
     
         const data = {quantity:quantity}
         const result = await Cart.updateOne({_id:id},data)
-            console.log(result.matchedCount)
-            res.status(200).send(result.matchedCount)
+        // console.log(result.matchedCount)
+        res.status(200).send(result.matchedCount)
         
     } catch (error) {
         console.log(error)
@@ -123,18 +130,28 @@ const decreaseCartItemQuantity = async(req,res)=>{
 const addToOrders = async (req,res)=>{
     const {userId} = req.query
     const cartItems = await Cart.find({userId:userId}).exec()
-    if(!cartItems){
+    if(!cartItems || cartItems.length === 0){
         return res.status(404).send("Could not be added")
     }
-    cartItems.forEach(async (item)=>{
-        const id = item._id
-        const product = item.product
-        const quantity = item.quantity
-        const userId = item.userId
-        const order = new Orders({product,quantity,userId})
+
+    for (const item of cartItems){
+        const order = new Orders({
+            product: item.product,
+            quantity: item.quantity,
+            userId:item.userId
+        })
         await order.save()
         await Cart.deleteOne({_id:id}).exec()
-    })
+    }
+    // cartItems.forEach(async (item)=>{
+    //     const id = item._id
+    //     const product = item.product
+    //     const quantity = item.quantity
+    //     const userId = item.userId
+    //     const order = new Orders({product,quantity,userId})
+    //     await order.save()
+    //     await Cart.deleteOne({_id:id}).exec()
+    // })
     res.status(200).send("All cart items added to order")
 }
